+++ source/components/panel/panel.js
@@ -75,7 +75,7 @@ var Panel = {
         } else if (typeof buttons === "object" && Utils.objectLength(buttons) > 0) {
             customButtons = buttons;
         } else {
-            console.log("Unknown format for custom buttons", buttons);
+            console.warn("Unknown format for custom buttons", buttons);
             return ;
         }
 
@@ -148,7 +148,7 @@ var Panel = {
 
             if (o.collapsible === true) {
                 var collapseToggle = $("<span>").addClass("dropdown-toggle marker-center active-toggle").addClass(o.clsCollapseToggle).appendTo(title);
-                element.collapse({
+                Metro.makePlugin(element, "collapse", {
                     toggleElement: collapseToggle,
                     duration: o.collapseDuration,
                     onCollapse: o.onCollapse,
@@ -176,7 +176,7 @@ var Panel = {
                 dragElement = panel;
             }
 
-            panel.draggable({
+            Metro.makePlugin(panel, "draggable", {
                 dragElement: dragElement,
                 onDragStart: o.onDragStart,
                 onDragStop: o.onDragStop,