+++ source/components/window/window.js
@@ -355,7 +355,7 @@ var Window = {
             resizer.appendTo(win);
             win.addClass("resizable");
 
-            win.resizable({
+            Metro.makePlugin(win, "resizable", {
                 minWidth: o.minWidth,
                 minHeight: o.minHeight,
                 maxWidth: o.maxWidth,
@@ -601,7 +601,7 @@ var Window = {
     toggleDraggable: function(){
         var element = this.element, win = this.win;
         var flag = JSON.parse(element.attr("data-draggable"));
-        var drag = win.data("draggable");
+        var drag = Metro.getPlugin(win, "draggable");
         if (flag === true) {
             drag.on();
         } else {
@@ -612,7 +612,7 @@ var Window = {
     toggleResizable: function(){
         var element = this.element, win = this.win;
         var flag = JSON.parse(element.attr("data-resizable"));
-        var resize = win.data("resizable");
+        var resize = Metro.getPlugin(win, "resizable");
         if (flag === true) {
             resize.on();
             win.find(".resize-element").removeClass("resize-element-disabled");
@@ -685,66 +685,53 @@ Metro['window'] = {
     },
 
     min: function(el, a){
-        var $$ = Utils.$();
         if (!this.isWindow(el)) {
             return false;
         }
-        var win = $$(el).data("window");
-        win.min(a);
+        Metro.getPlugin(el,"window").min(a);
     },
 
     max: function(el, a){
-        var $$ = Utils.$();
         if (!this.isWindow(el)) {
             return false;
         }
-        var win = $$(el).data("window");
-        win.max(a);
+        Metro.getPlugin(el, "window").max(a);
     },
 
     show: function(el){
-        var $$ = Utils.$();
         if (!this.isWindow(el)) {
             return false;
         }
-        var win = $$(el).data("window");
-        win.show();
+        Metro.getPlugin(el, "window").show();
     },
 
     hide: function(el){
-        var $$ = Utils.$();
         if (!this.isWindow(el)) {
             return false;
         }
-        var win = $$(el).data("window");
-        win.hide();
+        Metro.getPlugin(el, "window").hide();
     },
 
     toggle: function(el){
-        var $$ = Utils.$();
         if (!this.isWindow(el)) {
             return false;
         }
-        var win = $$(el).data("window");
-        win.toggle();
+        Metro.getPlugin(el, "window").toggle();
     },
 
     isOpen: function(el){
-        var $$ = Utils.$();
         if (!this.isWindow(el)) {
             return false;
         }
-        var win = $$(el).data("window");
+        var win = Metro.getPlugin(el,"window");
         return win.isOpen();
     },
 
     close: function(el){
-        var $$ = Utils.$();
         if (!this.isWindow(el)) {
             return false;
         }
-        var win = $$(el).data("window");
-        win.close();
+        Metro.getPlugin(el, "window").close();
     },
 
     create: function(options){