+++ docs/metro/js/metro.js
@@ -3596,7 +3596,7 @@ var isTouch = (('ontouchstart' in window) || (navigator["MaxTouchPoints"] > 0) |
 var Metro = {
 
     version: "4.3.1",
-    compileTime: "25/09/2019 13:47:16",
+    compileTime: "25/09/2019 14:20:02",
     buildNumber: "738",
     isTouchable: isTouch,
     fullScreenEnabled: document.fullscreenEnabled,
@@ -7856,7 +7856,7 @@ var Audio = {
 
         this.preloader = preloader;
 
-        streamSlider.slider({
+        Metro.makePlugin(streamSlider, "slider", {
             clsMarker: "bg-red",
             clsHint: "bg-cyan fg-white",
             clsComplete: "bg-cyan",
@@ -7880,7 +7880,7 @@ var Audio = {
             stream.hide();
         }
 
-        volumeSlider.slider({
+        Metro.makePlugin(volumeSlider, "slider", {
             clsMarker: "bg-red",
             clsHint: "bg-cyan fg-white",
             hint: true,
@@ -7912,7 +7912,7 @@ var Audio = {
 
         if (o.muted) {
             that.volumeBackup = audio.volume;
-            that.volume.data('slider').val(0);
+            Metro.getPlugin(that.volume, 'slider').val(0);
             audio.volume = 0;
         }
 
@@ -7944,7 +7944,7 @@ var Audio = {
         element.on("timeupdate", function(){
             var position = Math.round(audio.currentTime * 100 / that.duration);
             that._setInfo(audio.currentTime, that.duration);
-            that.stream.data('slider').val(position);
+            Metro.getPlugin(that.stream, 'slider').val(position);
             Utils.exec(o.onTime, [audio.currentTime, that.duration, audio, player], element[0]);
         });
 
@@ -7967,12 +7967,12 @@ var Audio = {
         });
 
         element.on("stop", function(){
-            that.stream.data('slider').val(0);
+            Metro.getPlugin(that.stream, 'slider').val(0);
             Utils.exec(o.onStop, [audio, player], element[0]);
         });
 
         element.on("ended", function(){
-            that.stream.data('slider').val(0);
+            Metro.getPlugin(that.stream, 'slider').val(0);
             Utils.exec(o.onEnd, [audio, player], element[0]);
         });
 
@@ -8016,12 +8016,11 @@ var Audio = {
         this.muted = !this.muted;
         if (this.muted === false) {
             this.audio.volume = this.volumeBackup;
-            this.volume.data('slider').val(this.volumeBackup * 100);
         } else {
             this.volumeBackup = this.audio.volume;
-            this.volume.data('slider').val(0);
             this.audio.volume = 0;
         }
+        Metro.getPlugin(this.volume, 'slider').val(this.muted === false ? this.volumeBackup * 100 : 0);
     },
 
     _setInfo: function(a, b){
@@ -8030,7 +8029,7 @@ var Audio = {
 
     _setBuffer: function(){
         var buffer = this.audio.buffered.length ? Math.round(Math.floor(this.audio.buffered.end(0)) / Math.floor(this.audio.duration) * 100) : 0;
-        this.stream.data('slider').buff(buffer);
+        Metro.getPlugin(this.stream, 'slider').buff(buffer);
     },
 
     _setVolume: function(){
@@ -8074,7 +8073,7 @@ var Audio = {
     stop: function(){
         this.audio.pause();
         this.audio.currentTime = 0;
-        this.stream.data('slider').val(0);
+        Metro.getPlugin(this.stream, 'slider').val(0);
     },
 
     volume: function(v){
@@ -8087,7 +8086,7 @@ var Audio = {
         }
 
         this.audio.volume = v;
-        this.volume.data('slider').val(v*100);
+        Metro.getPlugin(this.volume, 'slider').val(v*100);
     },
 
     loop: function(){
@@ -28850,7 +28849,6 @@ var TreeView = {
 
         $.each(nodes, function(){
             var node = $(this);
-            var childCount = 0;
             var caption, icon;
 
             caption = node.data("caption");
@@ -29808,7 +29806,7 @@ var Video = {
             infoBox.hide();
         }
 
-        streamSlider.slider({
+        Metro.makePlugin(streamSlider, "slider", {
             clsMarker: "bg-red",
             clsHint: "bg-cyan fg-white",
             clsComplete: "bg-cyan",
@@ -29832,7 +29830,7 @@ var Video = {
             stream.hide();
         }
 
-        volumeSlider.slider({
+        Metro.makePlugin(volumeSlider, "slider", {
             clsMarker: "bg-red",
             clsHint: "bg-cyan fg-white",
             hint: true,
@@ -30059,12 +30057,11 @@ var Video = {
         this.muted = !this.muted;
         if (this.muted === false) {
             this.video.volume = this.volumeBackup;
-            Metro.getPlugin(this.volume[0], 'slider').val(this.volumeBackup * 100);
         } else {
             this.volumeBackup = this.video.volume;
-            Metro.getPlugin(this.volume[0], 'slider').val(0);
             this.video.volume = 0;
         }
+        Metro.getPlugin(this.volume, 'slider').val(this.muted === false ? this.volumeBackup * 100 : 0);
     },
 
     _setInfo: function(a, b){
@@ -30073,7 +30070,7 @@ var Video = {
 
     _setBuffer: function(){
         var buffer = this.video.buffered.length ? Math.round(Math.floor(this.video.buffered.end(0)) / Math.floor(this.video.duration) * 100) : 0;
-        Metro.getPlugin(this.stream[0], 'slider').buff(buffer);
+        Metro.getPlugin(this.stream, 'slider').buff(buffer);
     },
 
     _setVolume: function(){