+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + [x] m4q: upd to 1.0.1
 + [x] m4q: fix initialization when `metro4` added into `head`
 + [x] Init: fix issue #1409 
++ [x] Dropdown: fix interop between Metro4 and jQuery  
 
 ### 4.3.0
 + [x] jQuery: Goodbye!!!